% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{aggregate_conmed_profile_tsv}
\alias{aggregate_conmed_profile_tsv}
\title{Merge all individual TSV files across studies into a single one for a given DrugBasket}
\usage{
aggregate_conmed_profile_tsv(
  input_tsv_path = NULL,
  tsv_pattern = NULL,
  basket_name = NULL,
  output_filename_prefix = NULL
)
}
\arguments{
\item{input_tsv_path}{directory from which stored results should be loaded}

\item{tsv_pattern}{file pattern identifying files to load}

\item{basket_name}{basket name to use in aggregated file name}

\item{output_filename_prefix}{prefix to use in aggregated file name}
}
\description{
Merge all individual TSV files across studies into a single one for a given DrugBasket
}
