% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{basket_demo}
\alias{basket_demo}
\title{Demo drug basket}
\format{
A tibble with 5 rows and 2 columns
}
\usage{
basket_demo
}
\description{
Classifications of some multi-agent medications into basket classes
}
\keyword{datasets}
