% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{complement_with_drug_free_pop}
\alias{complement_with_drug_free_pop}
\title{Complement profiles with the DRUG_FREE cohort}
\usage{
complement_with_drug_free_pop(
  cm_profiles = NULL,
  overall_pop = NULL,
  full_data = NULL,
  drug_free_label = "DRUG_FREE",
  treatments = c("trt1")
)
}
\arguments{
\item{cm_profiles}{concomitant medication profiles}

\item{overall_pop}{identifiers for the full population data}

\item{full_data}{full population data}

\item{drug_free_label}{label to use for the drug free population}

\item{treatments}{vector of treatment names as found in the exposure domain (to distinguish from control)}
}
\value{
output \code{data.frame} complemented with subjects who did not take any medications in the drug basket
}
\description{
Complement profiles with the DRUG_FREE cohort
}
