% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{compute_cm_overlaps}
\alias{compute_cm_overlaps}
\title{Compute overlapping classes}
\usage{
compute_cm_overlaps(cm_data = NULL, uncertain_basket_pattern = NULL)
}
\arguments{
\item{cm_data}{input \code{data.frame} of clinical data with concomitant medications}

\item{uncertain_basket_pattern}{regex to use to identify uncertain basket categories}
}
\value{
a \code{data.frame} of outputs - classifications of subjects into conmeds profiles and additional data
}
\description{
Compute overlapping classes
}
