% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConMedClassify-package.R
\docType{package}
\name{ConMedClassify-package}
\alias{ConMedClassify}
\alias{ConMedClassify-package}
\title{ConMedClassify: Concomitant Medication Profile Classifications}
\description{
Process clinical ADaM-format data and classify subjects
    into profiles based on their concomitant medication usage within 
    a clinical trial.
}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID})

Authors:
\itemize{
  \item Vincent Rouilly \email{rouilly.vincent@gene.com}
}

Other contributors:
\itemize{
  \item Genentech [copyright holder, funder]
}

}
\keyword{internal}
