% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_clinical_data}
\alias{demo_clinical_data}
\title{Demo clinical data}
\format{
A tibble with 4 rows and 11 columns with nested.CM, nested.TTE, and nested.EX
}
\usage{
demo_clinical_data
}
\description{
Nested CM, EX, and TTE domains, along with some demographics
}
\keyword{datasets}
