% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF.R
\name{DF_to_tibble}
\alias{DF_to_tibble}
\title{Convert a S4Vectors::DataFrame to a tibble}
\usage{
DF_to_tibble(d, add_rownames = FALSE)
}
\arguments{
\item{d}{input \code{data.frame} of clinical data}

\item{add_rownames}{(\code{logical}) should rownames be added?}
}
\value{
a \code{tibble} version of \code{d}
}
\description{
Convert a S4Vectors::DataFrame to a tibble
}
\details{
This function carefully deals with nested domains
}
