% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{ex_records}
\alias{ex_records}
\title{Extract records from nested.EX domain}
\usage{
ex_records(d = NULL)
}
\arguments{
\item{d}{input \code{data.frame} of clinical data with concomitant medications}
}
\value{
a \code{data.frame} of exposure records
}
\description{
Extract records from nested.EX domain
}
\details{
Additionally, filters out \code{AVISIT == OVERALL} records and 'not treated' subjects,
and records where \code{EXOCCUR != "Y"}.
}
