% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_basket_from_cm}
\alias{extract_basket_from_cm}
\title{Filter data to concomitant medication records matching a given basket}
\usage{
extract_basket_from_cm(
  nested_cm_data = NULL,
  drug_basket = NULL,
  unknownDoseCategoryLabel = "UNKDOSE"
)
}
\arguments{
\item{nested_cm_data}{input \code{data.frame} of clinical data with nested conmeds}

\item{drug_basket}{input \code{data.frame} of drug basket classifications}

\item{unknownDoseCategoryLabel}{suffix to use when dose is unknown}
}
\value{
a \code{data.frame} of concomitant medication data with classes merged
}
\description{
Filter data to concomitant medication records matching a given basket
}
\details{
Performs some cleanup of conmeds data.
}
