% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_conmed_profiles_to_tsv}
\alias{extract_conmed_profiles_to_tsv}
\title{Extract all drug baskets profiles for a given study to a TSV file}
\usage{
extract_conmed_profiles_to_tsv(
  d = NULL,
  drug_baskets = NULL,
  drug_basket_data_objects = NULL,
  output_dir = "outputs",
  output_filename_prefix = "ConMedProfiles_",
  treatments = c("trt1")
)
}
\arguments{
\item{d}{input \code{data.frame} of clinical data with concomitant medications}

\item{drug_baskets}{a \code{data.frame} of drug basket parameters, one per row. See details}

\item{drug_basket_data_objects}{a \code{list} of drug basket classifications, aligned to \code{drug_baskets}}

\item{output_dir}{directory to which results should be written. Should already exist}

\item{output_filename_prefix}{prefix to use for written files}

\item{treatments}{vector of treatment names as found in the exposure domain (to distinguish from control)}
}
\value{
nothing, used for the side-effect of writing results files
}
\description{
Extract all drug baskets profiles for a given study to a TSV file
}
\details{
\code{drug_baskets} should be a \code{data.frame} containing one row for each basket, with columns
"drug_basket_name", "combo_drug_label", "drug_free_label".

This package stores a \code{drug_basket} object for demonstration.
}
