% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{extract_data_for_cm_profiling}
\alias{extract_data_for_cm_profiling}
\title{Extract endpoint data and prepare for profiling}
\usage{
extract_data_for_cm_profiling(d = NULL, pop_of_interest = NULL)
}
\arguments{
\item{d}{input \code{data.frame} of clinical data}

\item{pop_of_interest}{input \code{data.frame} containing identifiers corresponding to the population of interest}
}
\value{
a \code{data.frame} containing endpoint data for the population of interest
}
\description{
Extract endpoint data and prepare for profiling
}
