% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF.R
\name{extract_paramcd}
\alias{extract_paramcd}
\title{Extract a PARAMCD-AVAL style key-value entries from nested data}
\usage{
extract_paramcd(d, nested.X, param.code, val.col, col.name = NA)
}
\arguments{
\item{d}{input \code{data.frame} of clinical data}

\item{nested.X}{nested column}

\item{param.code}{PARAMCD entry to which results should be filtered}

\item{val.col}{column containing the value to be extracted, corresponding to the PARAMCD}

\item{col.name}{column name to use for the extracted value}
}
\value{
a \code{data.frame} of extracted data
}
\description{
Extract a PARAMCD-AVAL style key-value entries from nested data
}
