% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{find_MissedAllDosePatients}
\alias{find_MissedAllDosePatients}
\title{Identify subjects who missed all doses of their treatment}
\usage{
find_MissedAllDosePatients(d)
}
\arguments{
\item{d}{input \code{data.frame} of clinical data with concomitant medications}
}
\value{
a \code{data.frame} of the filtered data
}
\description{
Identify subjects who missed all doses of their treatment
}
