% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{flag_uncertain_cm_records}
\alias{flag_uncertain_cm_records}
\title{Flag uncertain (timing) concomitant medication records}
\usage{
flag_uncertain_cm_records(cm_data = NULL)
}
\arguments{
\item{cm_data}{input \code{data.frame} of clinical data with concomitant medications}
}
\value{
a \code{data.frame} of concomitant medication data with an additional \code{UNCERTAIN_TIMING} column
}
\description{
Flag uncertain (timing) concomitant medication records
}
