% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{get_safety_pop}
\alias{get_safety_pop}
\title{Extract safety population}
\usage{
get_safety_pop(d)
}
\arguments{
\item{d}{input \code{data.frame} of clinical data}
}
\value{
a \code{data.frame} containing just the \code{STUDYID} and \code{USUBJID}
for the safety-evaluable population.
}
\description{
Filters input data to \verb{SAFFL == "Y}, the safety-evaluable population
}
