% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{identify_combo_and_uncertain_cm}
\alias{identify_combo_and_uncertain_cm}
\title{Flag uncertain timing and COMBO records}
\usage{
identify_combo_and_uncertain_cm(
  cm_data = NULL,
  combo_drug_label_prefix = "DRUG",
  combo_drug_label_suffix = "COMBO",
  combo_drug_uncertain_timing_suffix = "UNCERTAIN_TIMING",
  class_uncertain_timing_suffix = "UNCERTAIN_TIMING",
  separator = "_"
)
}
\arguments{
\item{cm_data}{input \code{data.frame} of clinical data with concomitant medications}

\item{combo_drug_label_prefix}{prefix to use for COMBO records}

\item{combo_drug_label_suffix}{suffix to use for COMBO records}

\item{combo_drug_uncertain_timing_suffix}{suffix to use for uncertain timing records}

\item{class_uncertain_timing_suffix}{class to assign uncertain timing records}

\item{separator}{separator to use bewteen prefix, class, and suffix}
}
\value{
a \code{data.frame} of concomitant medication records, with labels processed
}
\description{
Flag uncertain timing and COMBO records
}
