% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{impute_cm_days}
\alias{impute_cm_days}
\title{Impute end of concomitant medication exposure}
\usage{
impute_cm_days(cm_data = NULL, treatments = c("trt1"))
}
\arguments{
\item{cm_data}{input \code{data.frame} of clinical data with concomitant medications}

\item{treatments}{vector of treatment names as found in the exposure domain (to distinguish from control)}
}
\value{
a \code{data.frame} of concomitant medication data with imputed values
}
\description{
Impute end of concomitant medication exposure
}
\details{
In some cases, the end day of concomitant medication exposure (\code{CMEND}) is not avaialble, either
due to not being recorded, the administration being ongoing, or exit of the trial.

This function attempts to recover a usable end day where possible.
}
