% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{process_treatment_days}
\alias{process_treatment_days}
\title{process_treatment_days}
\usage{
process_treatment_days(cm_data = NULL, treatment_lasting_effect = 21)
}
\arguments{
\item{cm_data}{input \code{data.frame} of conmeds data}

\item{treatment_lasting_effect}{how long (in days) the treatment is expected to be
impactful following exposure}
}
\value{
a \code{data.frame} of the input data with treatment days annotated
}
\description{
process_treatment_days
}
