% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{remove_prior_cm_records}
\alias{remove_prior_cm_records}
\title{Filter out PRIOR concomitant medication records}
\usage{
remove_prior_cm_records(cm_data = NULL)
}
\arguments{
\item{cm_data}{input \code{data.frame} of clinical data with concomitant medications}
}
\value{
a \code{data.frame} of filtered concomitant medication data
}
\description{
Filter out PRIOR concomitant medication records
}
\details{
Removes records corresponding to concomitant medications taken prior to treatment
}
