% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF.R
\name{unnest_domain}
\alias{unnest_domain}
\title{Unnest a nested domain}
\usage{
unnest_domain(
  data,
  cols,
  .drop = NA,
  .id = NULL,
  .sep = NULL,
  .preserve = NULL
)
}
\arguments{
\item{data}{input clinical data with nested column(s)}

\item{cols}{column to unnest}

\item{.drop}{passed to \code{tidyr::unnest_legacy}}

\item{.id}{passed to \code{tidyr::unnest_legacy}}

\item{.sep}{passed to \code{tidyr::unnest_legacy}}

\item{.preserve}{passed to \code{tidyr::unnest_legacy}}
}
\value{
input data with the requested column unnested
}
\description{
Specific to this structure of nested data, but otherwise similar to the \code{tidyr} version
}
